+++ NEWS.md
@@ -1,3 +1,5 @@
+# cheapr (development version)
+
 # cheapr 1.0.1
 
 - R version 4.0.0 now required as also required by cpp11.